Add support for EditPersonDescriptor refactoring
